# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KEMRI-CGHR Influenza Program Clinical Data Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /anc:
    get:
      operationId: listAncVisits
      tags: [anc]
      summary: List ANC visits
      parameters:
        - name: screeningId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of ANC visits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AncVisit"
    post:
      operationId: createAncVisit
      tags: [anc]
      summary: Create an ANC visit record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AncVisitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AncVisit"
  /anc/{id}:
    get:
      operationId: getAncVisit
      tags: [anc]
      summary: Get an ANC visit record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ANC visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AncVisit"
        "404":
          description: Not found
    patch:
      operationId: updateAncVisit
      tags: [anc]
      summary: Update an ANC visit record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AncVisitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AncVisit"
    delete:
      operationId: deleteAncVisit
      tags: [anc]
      summary: Delete an ANC visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteInput"
      responses:
        "200":
          description: Deleted

  # Delivery
...
  /closeout:
    get:
      operationId: listCloseouts
      tags: [closeout]
      summary: List closeout records
      responses:
        "200":
          description: List of closeouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Closeout"
    post:
      operationId: createCloseout
      tags: [closeout]
      summary: Create a closeout record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseoutInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Closeout"
  /closeout/{screeningId}:
    get:
      operationId: getCloseout
      tags: [closeout]
      summary: Get a closeout record by screening ID
      parameters:
        - name: screeningId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Closeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Closeout"
        "404":
          description: Not found
    patch:
      operationId: updateCloseout
      tags: [closeout]
      summary: Update a closeout record
      parameters:
        - name: screeningId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseoutUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Closeout"
    delete:
      operationId: deleteCloseout
      tags: [closeout]
      summary: Delete a closeout record
      parameters:
        - name: screeningId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteInput"
      responses:
        "200":
          description: Deleted

  # Audit
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, fullName, initials, role]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        initials:
          type: string
        role:
          type: string
    DeleteInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # Screening
...
    AncVisit:
      type: object
      required: [id, screeningId, visitNumber, visitDate, createdBy, createdAt]
      properties:
        id:
          type: integer
        screeningId:
          type: string
        visitNumber:
          type: integer
        visitDate:
          type: string
        gestationalAge:
          type: ["integer", "null"]
        weight:
          type: ["number", "null"]
        bp:
          type: ["string", "null"]
        fundalHeight:
          type: ["number", "null"]
        muac:
          type: ["number", "null"]
        complaints:
          type: ["string", "null"]
        medication:
          type: ["string", "null"]
        nextAppointment:
          type: ["string", "null"]
        createdBy:
          type: string
        createdAt:
          type: string
    AncVisitInput:
      type: object
      required: [screeningId, visitNumber, visitDate]
      properties:
        screeningId:
          type: string
        visitNumber:
          type: integer
        visitDate:
          type: string
        gestationalAge:
          type: integer
        weight:
          type: number
        bp:
          type: string
        fundalHeight:
          type: number
        muac:
          type: number
        complaints:
          type: string
        medication:
          type: string
        nextAppointment:
          type: string
    AncVisitUpdate:
      type: object
      properties:
        visitDate:
          type: string
        gestationalAge:
          type: integer
        weight:
          type: number
        bp:
          type: string
        fundalHeight:
          type: number
        muac:
          type: number
        complaints:
          type: string
        medication:
          type: string
        nextAppointment:
          type: string
        reason:
          type: string

    # Delivery
...
    Closeout:
      type: object
      required: [id, screeningId, terminationDate, participantStatus, createdBy, createdAt]
      properties:
        id:
          type: integer
        screeningId:
          type: string
        terminationDate:
          type: string
        participantStatus:
          type: string
        discontinuationReason:
          type: ["string", "null"]
        createdBy:
          type: string
        createdAt:
          type: string
    CloseoutInput:
      type: object
      required: [screeningId, terminationDate, participantStatus]
      properties:
        screeningId:
          type: string
        terminationDate:
          type: string
        participantStatus:
          type: string
        discontinuationReason:
          type: string
    CloseoutUpdate:
      type: object
      properties:
        terminationDate:
          type: string
        participantStatus:
          type: string
        discontinuationReason:
          type: string
        reason:
          type: string

    # Audit
...