# lib/api-zod/src/generated/types/ancVisit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KEMRI-CGHR Influenza Program Clinical Data Management System
 * OpenAPI spec version: 0.1.0
 */

export interface AncVisit {
  id: number;
  screeningId: string;
  visitNumber: number;
  visitDate: string;
  /** @nullable */
  gestationalAge?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  bp?: string | null;
  /** @nullable */
  fundalHeight?: number | null;
  /** @nullable */
  muac?: number | null;
  /** @nullable */
  complaints?: string | null;
  /** @nullable */
  medication?: string | null;
  /** @nullable */
  nextAppointment?: string | null;
  createdBy: string;
  createdAt: string;
}