# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KEMRI-CGHR Influenza Program Clinical Data Management System
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  fullName: string;
  initials: string;
  role: string;
}